+++ docs/source/change_log.rst
@@ -17,6 +17,7 @@ Version 1.7.0
 * New features for client and server plugins
 * Add comparison of "trained" statistics to the campaign comparison
 * Support for including and importing Jinja templates from relative paths
+* New feature to import Campaigns from XML files.
 
 Version 1.6.0
 ^^^^^^^^^^^^^